src_async/tools_com/com_Material/com_soundspeaker.py
"""
Module com_soundspeaker - Fonctions de présence, fonctionnement et utilisation du haut-parleur
"""

import logging

logger = logging.getLogger(__name__)

def check_speaker_presence():
    """
    Vérifie la présence du haut-parleur.

    :return: True si le haut-parleur est présent, False sinon.
    :rtype: bool
    """
    try:
        # Vérification de la présence du haut-parleur ici
        presence = True  # Exemple de réponse (à adapter)
        if presence:
            logger.info("Speaker is present.")
        else:
            logger.info("Speaker is not present.")
        return presence
    except Exception:
        logger.error("Failed to check speaker presence.")
        return False

def start_speaker():
    """
    Démarre le haut-parleur pour la sortie audio.

    :return: True si le haut-parleur a démarré avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Démarrage du haut-parleur ici
        started = True  # Exemple de réponse (à adapter)
        if started:
            logger.info("Speaker started.")
        else:
            logger.error("Failed to start speaker.")
        return started
    except Exception:
        logger.error("Failed to start speaker.")
        return False

def play_audio(audio_file_path):
    """
    Joue un fichier audio à partir du haut-parleur.

    :param audio_file_path: Chemin du fichier audio à jouer.
    :type audio_file_path: str
    :return: True si l'audio a été joué avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Lecture de l'audio à partir du haut-parleur
        # Utilisation du fichier audio spécifié par audio_file_path
        played = True  # Exemple de réponse (à adapter)
        if played:
            logger.info("Audio played from speaker.")
        else:
            logger.error("Failed to play audio from speaker.")
        return played
    except Exception:
        logger.error("Failed to play audio from speaker.")
        return False

def stop_speaker():
    """
    Arrête la sortie audio du haut-parleur.

    :return: True si le haut-parleur a été arrêté avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Arrêt de la sortie audio du haut-parleur ici
        stopped = True  # Exemple de réponse (à adapter)
        if stopped:
            logger.info("Speaker stopped.")
        else:
            logger.error("Failed to stop speaker.")
        return stopped
    except Exception:
        logger.error("Failed to stop speaker.")
        return False
